add useAccountId and useAccountVisibility hooks to make things more flexible
