+++ app/javascript/mastodon/components/remote_hint.tsx
@@ -14,10 +14,15 @@ export const RemoteHint: React.FC<RemoteHintProps> = ({ accountId }) => {
   const account = useAppSelector((state) =>
     accountId ? state.accounts.get(accountId) : undefined,
   );
-  if (!account || !account.url || account.acct !== account.username) {
+  const domain = account?.acct ? account.acct.split('@')[1] : undefined;
+  if (
+    !account ||
+    !account.url ||
+    account.acct !== account.username ||
+    !domain
+  ) {
     return null;
   }
-  const domain = account.acct ? account.acct.split('@')[1] : undefined;
 
   return (
     <TimelineHint
